requirements.txt
streamlit
torch
torchvision
pillow
numpy
opencv-python
git+https://github.com/jacobgil/pytorch-grad-cam.git